+++ src/vs/workbench/contrib/accessibility/browser/accessibility.contribution.ts
@@ -18,7 +18,7 @@ import { SpeechAccessibilitySignalContribution } from 'vs/workbench/contrib/spee
 import { AccessibleViewInformationService, IAccessibleViewInformationService } from 'vs/workbench/services/accessibility/common/accessibleViewInformationService';
 import { IAccessibleViewService } from 'vs/platform/accessibility/browser/accessibleView';
 import { AccessibleViewService } from 'vs/workbench/contrib/accessibility/browser/accessibleView';
-import { CommentAccessibleViewContribution, InlineCompletionsAccessibleViewContribution, AccesibleViewHelpContribution, AccesibleViewContributions } from 'vs/workbench/contrib/accessibility/browser/accessibleViewContributions';
+import { InlineCompletionsAccessibleViewContribution, AccesibleViewHelpContribution, AccesibleViewContributions } from 'vs/workbench/contrib/accessibility/browser/accessibleViewContributions';
 import { ExtensionAccessibilityHelpDialogContribution } from 'vs/workbench/contrib/accessibility/browser/extensionAccesibilityHelp.contribution';
 
 registerAccessibilityConfiguration();
@@ -32,7 +32,6 @@ workbenchRegistry.registerWorkbenchContribution(UnfocusedViewDimmingContribution
 
 workbenchRegistry.registerWorkbenchContribution(AccesibleViewHelpContribution, LifecyclePhase.Eventually);
 workbenchRegistry.registerWorkbenchContribution(AccesibleViewContributions, LifecyclePhase.Eventually);
-workbenchRegistry.registerWorkbenchContribution(CommentAccessibleViewContribution, LifecyclePhase.Eventually);
 workbenchRegistry.registerWorkbenchContribution(InlineCompletionsAccessibleViewContribution, LifecyclePhase.Eventually);
 
 registerWorkbenchContribution2(AccessibilityStatus.ID, AccessibilityStatus, WorkbenchPhase.BlockRestore);
